+++ iam.tf
@@ -27,3 +27,17 @@ resource "google_kms_crypto_key_iam_binding" "auto_unseal_iam_binding" {
     "serviceAccount:${google_service_account.vault_sa.email}",
   ]
 }
+
+
+# add Cloud Run Admin and Service Account User roles to Cloud Build service account
+resource "google_project_iam_member" "iam_run_admin" {
+  project = var.project_id
+  role    = "roles/run.admin"
+  member  = "serviceAccount:${var.project_number}@cloudbuild.gserviceaccount.com"
+}
+
+resource "google_project_iam_member" "iam_sa_user" {
+  project = var.project_id
+  role    = "roles/iam.serviceAccountUser"
+  member  = "serviceAccount:${var.project_number}@cloudbuild.gserviceaccount.com"
+}