+++ provider.tf
@@ -1,13 +1,13 @@
 terraform {
-  required_version = "~> 0.15.0"
+  required_version = "~> 1.0"
   required_providers {
     google = {
       source  = "hashicorp/google"
-      version = "~> 3.0"
+      version = "~> 4.0"
     }
     google-beta = {
       source  = "hashicorp/google-beta"
-      version = "~> 3.0"
+      version = "~> 4.0"
     }
   }
 }