+++ variables.tf
@@ -1,6 +1,8 @@
 #-----------------------
 # provider variables
 #-----------------------
+variable "project_number" {}
+
 variable "project_id" {}
 
 variable "credentials_file_path" {}
@@ -33,6 +35,11 @@ variable "location" {
   default     = "US"
 }
 
+variable "force_destroy" {
+  description = "Force destroy the GCS bucket even if it's not empty"
+  default     = "true"
+}
+
 
 #-------------------------
 # Secret Manager
@@ -41,4 +48,29 @@ variable "location" {
 variable "vault_server_config" {
   description = "Path of Vault server config"
   default     = "./cloud-run/vault-server.hcl"
+}
+
+
+#-------------------------
+# Cloud Build
+#-------------------------
+
+variable "github_repo_owner" {
+  description = "GitHub repo owner name (user or organization)"
+}
+
+variable "github_repo_name" {
+  description = "GitHub repo name.  Not the full URI"
+}
+
+variable "cloudrun_service_name" {}
+
+variable "cloudrun_region" {
+  description = "Region where Cloud Run is deployed.  This may be different from the region due to the some feastures (i.e. custom domains) that are still in beta and not support everywehre yet."
+  default     = "us-east4"
+}
+
+variable "trigger_approval_required" {
+  description = "Require approval before build executes"
+  default     = "true"
 }